Remove mouse support for terminal
Having to use the mouse on the terminal is completely wrong - I reckon that it's best for everyone just to learn how to use keyboard well enough to be able to move around the document.

At the end of the day it's faster.

If anyone is against this, please let me know.